generic dice refactoring
