+++ src/components/generic-dice/animate-dice.scss
